// src/data/models/setting/response/profile_griya.ts
type ProfileGriya = {
  id: number;
  name: string;
  email: string;
  phoneNumber: string;
  logo?: string; // Optional field
  address: string;
  about: string;
  vision: string;
  mission: string;
  createdAt: string; // ISO date format
  updatedAt: string; // ISO date format
};

export default ProfileGriya;